63-theme: remove theme and table, create components package.
