+++ packages/components/.eslintrc.json
@@ -10,6 +10,5 @@
     "no-console": "off",
     "react/react-in-jsx-scope": "off"
   }
-
 }
