+++ packages/components/README.md
@@ -1,0 +1,1 @@
+# STC Component Library
