+++ packages/components/src/index.ts
@@ -1,0 +1,10 @@
+import './index.css'
+
+// Components
+export type { TableProps } from './table'
+export { Table } from './table'
+
+
+// Styles
+export { StcStylesProvider, useStyles, getStyles, themeConfig } from './styles'
+export type { StclandStyles, TableStyles } from './styles'
