+++ packages/components/src/styles/index.ts
@@ -1,0 +1,6 @@
+
+import '../index.css'
+
+export { StcStylesProvider, useStyles, getStyles, themeConfig } from './StylesProvider'
+
+export type { StclandStyles, TableStyles } from './StylesProvider'
