+++ packages/components/src/table/index.ts
@@ -1,0 +1,2 @@
+export type { Props as TableProps } from './Table'
+export { default as Table } from './Table'
