+++ packages/components/src/table/withStyles/Table.tsx
@@ -1,15 +1,15 @@
-import '../../index.css'
-import { useStyles } from '@stcland/theme'
-import { TableProps } from '@/index'
+import type { TableProps } from '..'
 import Table  from '../Table'
+import { useStyles } from '../../styles/StylesProvider'
 
 //*****************************************************************************
   // Components
 //*****************************************************************************
 
 const TableWithStyles = ( props: TableProps) => {
+  const { styles }  = useStyles()
 
-  const { styles } = useStyles()
+  console.log('IN TABLE WRAPPER------------', styles)
 
   return (
     <Table {...props} customStyles={styles?.table}/>
