+++ packages/components/src/withStyles/index.ts
@@ -1,0 +1,2 @@
+// All component withStyles exports.
+export { Table } from '../table/withStyles'
