+++ packages/components/tsconfig.json
@@ -9,13 +9,11 @@
     "allowArbitraryExtensions": true,
     "baseUrl": ".",
     "paths": {
-       "@/*": ["src/*"],
-      "@storybook/*": ["../storybook*"]
+       "@/*": ["src"],
     }
   },
   "include": [
-    "src", "./index.ts",
-    "./src/components/withStyles/index.ts"
+    "src",
   ],
   "exclude": [
     "node_modules"
