--- packages/table/.lintstagedrc.js
@@ -1,8 +1,0 @@
-import baseConfig from '../config/.lintstagedrc.base.js'
-
-export default {
-  ...baseConfig,
-  '*.{ts,tsx}': [
-    () => 'pnpm ts-check',
-  ],
-}
