--- packages/table/CHANGELOG.md
@@ -1,36 +1,0 @@
-# @stcland/table
-
-## 0.0.6
-
-### Patch Changes
-
-- Update context to StyleProvider
-- Updated dependencies
-  - @stcland/theme@0.0.3
-
-## 0.0.5
-
-### Patch Changes
-
-- Updated dependencies
-  - @stcland/theme@0.0.2
-
-## 0.0.4
-
-### Patch Changes
-
-- Added theme package as source of truth for tailwind
-- Updated dependencies
-  - @stcland/theme@0.0.1
-
-## 0.0.3
-
-### Patch Changes
-
-- latest version.
-
-## 0.0.2
-
-### Patch Changes
-
-- Vite lib mode bundling css in js
