--- packages/table/src/index.css
@@ -1,1 +1,0 @@
-@import "@stcland/theme/dist/style.css"
