--- packages/table/src/index.ts
@@ -1,7 +1,0 @@
-import './index.css'
-
-export type { Props as TableProps } from './components/Table'
-
-export { default as Table } from './components/Table'
-
-// export {default as Table} from './components/withTheme'
