--- packages/table/tailwind.config.ts
@@ -1,8 +1,0 @@
-import { themeConfig } from '@stcland/theme'
-
-const config = {
-  ...themeConfig,
-  content: ['./src/**/*.tsx'],
-}
-
-export default config
