--- packages/theme/.eslintrc.json
@@ -1,15 +1,0 @@
-{
-  "root": true,
-  "extends": ["../config/eslint.react.base.json"],
-  "env": {
-    "node": true,
-    "es6": true
-  },
-  "ignorePatterns": ["dist", ".eslintrc.json"],
-   "rules": {
-    "no-console": "off",
-    "react/react-in-jsx-scope": "off"
-  }
-
-}
-
