--- packages/theme/CHANGELOG.md
@@ -1,19 +1,0 @@
-# @stcland/theme
-
-## 0.0.3
-
-### Patch Changes
-
-- Update context to StyleProvider
-
-## 0.0.2
-
-### Patch Changes
-
-- Remove src files for publishing
-
-## 0.0.1
-
-### Patch Changes
-
-- Added theme package as source of truth for tailwind
