--- packages/theme/src/index.ts
@@ -1,6 +1,0 @@
-
-import './global.css'
-
-export { StcStylesProvider, useStyles, themeConfig } from './StylesProvider'
-
-export type { StclandStyles, TableStyles } from './StylesProvider'
