--- packages/theme/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "extends": "../config/tsconfig.base.json",
-    "compilerOptions": {
-      "outDir": "dist",
-      "declaration": true,
-      "declarationMap": true,
-      "jsx": "react-jsx",
-      "allowImportingTsExtensions": true,
-      "allowArbitraryExtensions": true,
-      "baseUrl": ".",
-      "paths": {
-         "@/*": ["src"],
-      }
-    },
-    "include": [
-      "src",
-    ],
-    "exclude": [
-      "node_modules"
-    ]
-  }
-
