--- packages/theme/vite.config.ts
@@ -1,30 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react-swc'
-import path from 'path'
-import dts from 'vite-plugin-dts'
-
-export default defineConfig({
-  plugins: [
-    react(),
-    dts({ rollupTypes: true, insertTypesEntry: true,  }),
-  ],
-  build: {
-    outDir: './dist',
-    lib: {
-      entry: path.resolve(__dirname, 'src/index.ts'),
-      name: '@stcland/theme',
-      fileName: (format) => `index.${format}.js`,
-    },
-    rollupOptions: {
-      external: ['react', 'react-dom', 'tailwindcss'],
-      output: {
-        globals: {
-          react: 'React',
-          'react-dom': 'ReactDOM',
-        },
-      },
-    },
-    target: 'esnext',
-    sourcemap: true,
-  },
-})
