+++ packages/utils/src/classnames.ts
@@ -1,5 +1,5 @@
 import clsx, { ClassValue } from 'clsx'
-import { toPairs } from 'ramda'
+import { toPairs, mergeDeepWith, is } from 'ramda'
 import { twMerge } from 'tailwind-merge'
 
 /**
@@ -7,6 +7,11 @@ import { twMerge } from 'tailwind-merge'
  */
 export const cns = (...inputs: ClassValue[]): string => {
   return twMerge(clsx(inputs))
+}
+
+// Helper function to merge styles using classnames
+const mergeFn = (base: any, custom: any) => {
+  return is(String, base) && is(String, custom) ? cns(base, custom) : custom
 }
 
 /**
@@ -18,11 +23,21 @@ type WithCustomStyles = <CustomStylesShape extends Record<string, ClassValue>>(
   customStyles?: Partial<CustomStylesShape>
 ) => CustomStylesShape
 
+
 export const withCustomStyles: WithCustomStyles = (baseStyles, customStyles = {}) => {
   const newStyles : Record<string, ClassValue> = {}
   for (const [key, value] of toPairs(baseStyles as Record<string, any>))
     newStyles[key] = cns(value, customStyles?.[key])
   return newStyles as typeof baseStyles
+}
+
+
+/**
+ * Deep merges `baseStyles` and `customStyles`.
+ * Utilizes Ramda's `mergeDeepWith` for nested object merging.
+ */
+export const mergeCustomStyles: WithCustomStyles = (baseStyles, customStyles = {}) => {
+  return mergeDeepWith(mergeFn, baseStyles, customStyles) as typeof baseStyles
 }
 
 export default cns
